refactoring : moved all hearthstone data classes in a new module
