+++ hearthstone/src/main/java/net/hearthstats/core/GameMode.java
