+++ hearthstone/src/main/java/net/hearthstats/core/HeroClass.java
