+++ hearthstone/src/main/java/net/hearthstats/core/MatchOutcome.java
