+++ hearthstone/src/main/java/net/hearthstats/core/Rank.java
