+++ hearthstone/src/main/java/net/hearthstats/game/CardEventType.java
