+++ hearthstone/src/main/resources/net/hearthstats/core/AllSets.json
