+++ hearthstone/src/main/scala/net/hearthstats/core/Card.scala
