+++ hearthstone/src/main/scala/net/hearthstats/core/CardData.scala
