+++ hearthstone/src/main/scala/net/hearthstats/core/Deck.scala
