+++ hearthstone/src/main/scala/net/hearthstats/core/GameLog.scala
