+++ hearthstone/src/main/scala/net/hearthstats/core/HearthstoneMatch.scala
