+++ hearthstone/src/test/scala/net/hearthstats/core/CardParserJsonSpec.scala
